fix(truelayer): Skip unsupported account types
Some providers (e.g. barclaycard) do not support both "accounts" and
"cards". If the truelayer importer comes across such a provider, it will
fail to fetch the unsupported account type, and omit that from the
config

If the key is omitted from the config, we'll try to read from an invalid
key when we later look for all transactions